requirements.txt
transformers
torch
sentencepiece
pandas
requests
bert-score
matplotlib
python-dotenv